[mlir][transform] Expensive checks: look for unmapped ops in cache
As part of the "expensive checks", the transform dialect interpreter
maintain a cache of payload ops to payload op names. This is to detect
cases an op was deleted (but not removed from the mappings and the
cache) and a new op of different type was allocated at the same pointer
location.

This commit improves the expensive checks to look for ops that are in
the cache but that are not mapped to any handle. Only mapped ops should
be in the cache. The presence of unmapped ops can indicate incorrect
side effects (about handle consumptions etc.).

This change should make issues such as #72931 easier to debug.